POD AutoM/backend/requirements.txt
# POD AutoM Backend - Requirements
# Python 3.11+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
supabase==2.10.0

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Environment & Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Image Processing (for mockups)
Pillow==11.0.0

# OpenAI (for GPT Image)
openai==1.54.0

# Shopify
ShopifyAPI==12.6.0

# Pinterest (REST API - we'll use httpx)
# No official SDK, using httpx

# Utilities
python-dateutil==2.9.0
orjson==3.10.11

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
